refactor!: Move CDVAllowList out of public API
This isn't really accessible outside of implementation details of the
internal CDVIntentAndNavigationFilter plugin, so it can be a
project-level internal header that isn't exposed publicly.
