+++ CordovaLib/Classes/Private/Plugins/CDVIntentAndNavigationFilter/CDVAllowList.h
