+++ CordovaLib/include/Cordova/CDV.h
@@ -32,7 +32,6 @@
 #import <Cordova/CDVWebViewProcessPoolFactory.h>
 #import <Cordova/NSDictionary+CordovaPreferences.h>
 #import <Cordova/NSMutableArray+QueueAdditions.h>
-#import <Cordova/CDVAllowList.h>
 #import <Cordova/CDVScreenOrientationDelegate.h>
 #import <Cordova/CDVTimer.h>
 #import <Cordova/CDVURLSchemeHandler.h>
